sparql: Extract HIR types to new mz-hir crate
Extract `HirRelationExpr`, `HirScalarExpr`, and related HIR types from
`mz-sql` into a new `mz-hir` crate. This breaks the dependency of
`mz-sparql` on `mz-sql`, enabling `mz-sql` to eventually depend on
`mz-sparql` for direct SPARQL-to-HIR compilation without deferred
adapter-side processing.

The new `mz-hir` crate contains:
- All HIR type definitions (HirRelationExpr, HirScalarExpr, AggregateExpr,
  AggregateFunc, JoinKind, ColumnRef, WindowExpr, CoercibleScalarExpr, etc.)
- Pure builder methods and visitor implementations
- Display impls for HirScalarExpr and AggregateExpr
- ScalarOps impl for HirScalarExpr
- Explain impl for HirRelationExpr (with normalize_subqueries)
- DisplayText impl for HirRelationExpr (explain text formatting)

Methods that depend on mz-sql internals (type coercion, parameter binding,
expression simplification via lowering) remain in mz-sql as extension traits:
- CoercibleScalarExprExt (type_as, cast_to)
- HirRelationExprExt (bind_parameters, finish_maintained)
- HirScalarExprExt (bind_parameters, into_literal_int64, etc.)
- HirRelationExprLowering, HirScalarExprLowering (HIR-to-MIR lowering)
- HirScalarExprTransform (HIR transformations)

mz-sparql now depends on mz-hir instead of mz-sql for HIR types.
