+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
     "src/expr-derive",
     "src/expr-derive-impl",
     "src/expr-parser",
+    "src/hir",
     "src/expr-test-util",
     "src/fivetran-destination",
     "src/foundationdb",
@@ -168,6 +169,7 @@ default-members = [
     "src/expr-derive",
     "src/expr-derive-impl",
     "src/expr-parser",
+    "src/hir",
     "src/expr-test-util",
     "src/foundationdb",
     "src/frontegg-auth",
