+++ src/adapter/src/optimize/copy_to.rs
@@ -23,6 +23,7 @@ use mz_repr::explain::trace_plan;
 use mz_repr::{GlobalId, Timestamp};
 use mz_sql::optimizer_metrics::OptimizerMetrics;
 use mz_sql::plan::HirRelationExpr;
+use mz_sql::plan::HirRelationExprLowering;
 use mz_sql::session::metadata::SessionMetadata;
 use mz_storage_types::connections::Connection;
 use mz_storage_types::sinks::S3UploadInfo;
