+++ src/hir/src/explain/text.rs
@@ -23,35 +23,16 @@ use std::fmt;
 use mz_expr::Id;
 use mz_expr::explain::{HumanizedExplain, HumanizerMode, fmt_text_constant_rows};
 use mz_expr::virtual_syntax::{AlgExcept, Except};
-use mz_ore::error::ErrorExt;
 use mz_ore::str::{IndentLike, separated};
+use mz_repr::Datum;
 use mz_repr::Diff;
 use mz_repr::explain::{CompactScalarSeq, Indices, PlanRenderingContext};
 
-use crate::plan::hir::HirScalarExprExt;
-use crate::plan::{Hir, HirRelationExpr, JoinKind};
+use mz_repr::explain::text::DisplayText;
 
-pub(crate) trait HirRelationExprTextExplain {
-    fn fmt_text(
-        &self,
-        f: &mut fmt::Formatter<'_>,
-        ctx: &mut PlanRenderingContext<'_, HirRelationExpr>,
-    ) -> fmt::Result;
+use crate::{Hir, HirRelationExpr, JoinKind};
 
-    fn fmt_virtual_syntax(
-        &self,
-        f: &mut fmt::Formatter<'_>,
-        ctx: &mut PlanRenderingContext<'_, HirRelationExpr>,
-    ) -> fmt::Result;
-
-    fn fmt_raw_syntax(
-        &self,
-        f: &mut fmt::Formatter<'_>,
-        ctx: &mut PlanRenderingContext<'_, HirRelationExpr>,
-    ) -> fmt::Result;
-}
-
-impl HirRelationExprTextExplain for HirRelationExpr {
+impl DisplayText<PlanRenderingContext<'_, HirRelationExpr>> for HirRelationExpr {
     fn fmt_text(
         &self,
         f: &mut fmt::Formatter<'_>,
@@ -63,8 +44,10 @@ impl HirRelationExprTextExplain for HirRelationExpr {
             self.fmt_virtual_syntax(f, ctx)
         }
     }
+}
 
-    fn fmt_virtual_syntax(
+impl HirRelationExpr {
+    pub fn fmt_virtual_syntax(
         &self,
         f: &mut fmt::Formatter<'_>,
         ctx: &mut PlanRenderingContext<'_, HirRelationExpr>,
@@ -88,7 +71,7 @@ impl HirRelationExprTextExplain for HirRelationExpr {
         Ok(())
     }
 
-    fn fmt_raw_syntax(
+    pub fn fmt_raw_syntax(
         &self,
         f: &mut fmt::Formatter<'_>,
         ctx: &mut PlanRenderingContext<'_, HirRelationExpr>,
@@ -265,24 +248,10 @@ impl HirRelationExprTextExplain for HirRelationExpr {
                     write!(f, " limit={}", limit)?;
                 }
                 // We only print the offset if it is not trivial, i.e., not 0.
-                let offset_literal = offset.clone().try_into_literal_int64();
-                if !offset_literal.as_ref().is_ok_and(|&offset| offset == 0) {
-                    let offset = if offset.contains_parameters() {
-                        // If we are still before parameter binding, then we can't reduce it to a
-                        // literal, so just print the expression.
-                        // (Untested as of writing this, because we don't have an EXPLAIN that would
-                        // show the plan before parameter binding.)
-                        offset.to_string()
-                    } else {
-                        match offset_literal {
-                            Ok(offset) => offset.to_string(),
-                            // The following is also untested, because we error out in planning
-                            // if reducing the OFFSET results in an error.
-                            Err(err) => err.to_string_with_causes(),
-                        }
-                    };
+                let is_zero_offset = offset.as_literal().map_or(false, |d| d == Datum::Int64(0));
+                if !is_zero_offset {
                     write!(f, " offset={}", offset)?;
-                };
+                }
                 if let Some(expected_group_size) = expected_group_size {
                     write!(f, " exp_group_size={}", expected_group_size)?;
                 }
